Move reference test implementations to separate folders, move mock-browser-api to helpers
