+++ .eslintignore
@@ -1,2 +1,2 @@
 shared/js/content-scope/sjcl.js
-unit-test/background/reference-tests/
+unit-test/data/reference-tests/
