+++ unit-test/background/click-to-load-tests.js
@@ -1,4 +1,4 @@
-require('../background/mock-browser-api')
+require('../helpers/mock-browser-api')
 const chromeWrapper = require('../../shared/js/background/wrapper.es6.js')
 const tds = require('../../shared/js/background/trackers.es6')
 const tdsStorage = require('../../shared/js/background/storage/tds.es6')
