+++ unit-test/background/https.es6.js
@@ -1,4 +1,4 @@
-require('../background/mock-browser-api')
+require('../helpers/mock-browser-api')
 const testDomains = require('./../data/httpsTestDomains.json')
 const https = require('../../shared/js/background/https.es6')
 const httpsStorage = require('../../shared/js/background/storage/https.es6')
