+++ unit-test/background/reference-tests/privacy-configuration-tests.js
@@ -1,17 +1,17 @@
-require('../background/mock-browser-api')
+require('../../helpers/mock-browser-api')
 
-const tdsStorageStub = require('./../helpers/tds.es6')
-const tdsStorage = require('../../shared/js/background/storage/tds.es6')
+const tdsStorageStub = require('../../helpers/tds.es6')
+const tdsStorage = require('../../../shared/js/background/storage/tds.es6')
 
-const Site = require('../../shared/js/background/classes/site.es6')
-const utils = require('../../shared/js/background/utils.es6')
-const browserWrapper = require('../../shared/js/background/wrapper.es6')
+const Site = require('../../../shared/js/background/classes/site.es6')
+const utils = require('../../../shared/js/background/utils.es6')
+const browserWrapper = require('../../../shared/js/background/wrapper.es6')
 
-const contentScriptUtils = require('../../shared/js/content-scope/utils.js')
+const contentScriptUtils = require('../../../shared/js/content-scope/utils.js')
 
-const testSets = require('./reference-tests/privacy-configuration/tests.json')
+const testSets = require('../../data/reference-tests/privacy-configuration/tests.json')
 const configs = {
-    // FILE_MAP ./reference-tests/privacy-configuration/config*_reference.json
+    // FILE_MAP ../../data/reference-tests/privacy-configuration/config*_reference.json
 }
 
 const EXT_ID = 'ogigmfedpbpnnbcpgjloacccaibkaoip'
