+++ unit-test/background/reference-tests/tracker-reference-tests.js
@@ -1,12 +1,12 @@
 const fs = require('fs')
 const path = require('path')
 
-const tds = require('../../shared/js/background/trackers.es6')
-const tdsStorageStub = require('./../helpers/tds.es6')
+const tds = require('../../../shared/js/background/trackers.es6')
+const tdsStorageStub = require('../../helpers/tds.es6')
 
-const refTrackers = require('./reference-tests/tracker-radar-tests/TR-domain-matching/tracker_radar_reference.json')
-const refTests = require('./reference-tests/tracker-radar-tests/TR-domain-matching/domain_matching_tests.json')
-const refSurrogates = fs.readFileSync(path.join(__dirname, './reference-tests/tracker-radar-tests/TR-domain-matching/surrogates.txt'), 'utf8')
+const refTrackers = require('../../data/reference-tests/tracker-radar-tests/TR-domain-matching/tracker_radar_reference.json')
+const refTests = require('../../data/reference-tests/tracker-radar-tests/TR-domain-matching/domain_matching_tests.json')
+const refSurrogates = fs.readFileSync(path.join(__dirname, '../../data/reference-tests/tracker-radar-tests/TR-domain-matching/surrogates.txt'), 'utf8')
 
 describe('Tracker reference tests:', () => {
     beforeAll(() => {
