+++ unit-test/helpers/mock-browser-api.js
